SCMOD-7132: tests related to workflow-control moved to another class
